PM
Report function : senders and recipients well displayed.
recipients selected by a select button-> no mistake
A button to send a message to all managers (currently sending to all users of the Bu)
Changes made to start fitingt expectations about staff/management.

+++ rms/application/views/pm/list.php
@@ -1,17 +1,28 @@
+<? if($userlevel == 0){?>
+	</div>	
+		<div data-role="content" data-theme="a">
+<? } ?>
 			<?php if(count($messages)>0):?>
-				<table data-role="table" id="table-custom-2" data-insert="true" data-mode="reflow" data-filter="true" class="ui-body-d ui-shadow table-stripe ui-responsive" data-column-popup-theme="a" data-filter-placeholder="Filter message">
+				<table data-role="table" id="table-custom-2" data-insert="true" data-mode="reflow" class="ui-body-d ui-shadow table-stripe ui-responsive" data-column-popup-theme="a" <? if($userlevel >= 1){?>data-filter="true" data-filter-placeholder="Filter message"<?}?>>
 					<thead>
+						<tr>
 						<th><?php if($type != MSG_SENT) echo 'From'; else echo 'Recipients'; ?></th>
 						<th>Subject</th>
 						<th>Date</th>
-						<th>Reply</th>
-						<th><?php if($type != MSG_DELETED) echo 'Delete'; else echo 'Restore'; ?></th>
-						</tr>
+						<? if($userlevel >= 1){?>
+							<? if($type != MSG_SENT){ ?>
+								<th>Reply</th>
+							<? }else{ ?>
+								<th></th>
+							<? } ?>
+							<th><?php if($type != MSG_DELETED) echo 'Delete'; else echo 'Restore'; ?></th>
+							</tr>
+						<?}?>
 					</thead>
 					<tbody>
 						<?php for ($i=0; $i<count($messages); $i++): ?>
 						<tr>
-							<td>
+							<th>
 								<?php
 								if($type != MSG_SENT) echo $messages[$i][TF_PM_AUTHOR];
 								else
@@ -20,25 +31,29 @@
 									foreach ($recipients as $recipient)
 										echo (next($recipients)) ? $recipient.', ' : $recipient;
 								}?>
-							</td>
-							<td><a data-ajax="false" href='<?php echo site_url().'/pm/message/'.$messages[$i][TF_PM_ID]; ?>'><?php echo $messages[$i][TF_PM_SUBJECT] ?></a></td>
-							<td><?php echo $messages[$i][TF_PM_DATE]; ?></td>
-							<?php if($type != MSG_SENT): ?>
-								<td><?php echo '<a data-ajax="false" href="'.site_url().'/pm/send/'.$messages[$i][TF_PM_AUTHOR].'/RE&#58;'.$messages[$i][TF_PM_SUBJECT].'"> reply </a>' ?></td>
-							<?php endif; ?>
-							<td>
-								<?php if($type != MSG_DELETED)
-									echo '<a data-ajax="false" href="'.site_url().'/pm/delete/'.$messages[$i][TF_PM_ID].'/'.$type.'"> x </a>';
-								else
-									echo '<a data-ajax="false" href="'.site_url().'/pm/restore/'.$messages[$i][TF_PM_ID].'"> o </a>'; ?>
-								</td>
+							</th>
+							<th><a data-ajax="false" href='<?php echo site_url().'/pm/message/'.$messages[$i][TF_PM_ID]; ?>'><?php echo $messages[$i][TF_PM_SUBJECT] ?></a></th>
+							<th><?php echo $messages[$i][TF_PM_DATE]; ?></th>
+							<? if($userlevel >= 1){?>
+								<? if($type != MSG_SENT){ ?>
+									<th><?php echo '<a data-ajax="false" href="'.site_url().'/pm/send/'.$messages[$i][TF_PM_AUTHOR].'/RE&#58;'.$messages[$i][TF_PM_SUBJECT].'"> reply </a>' ?></th>
+								<? }else{ ?>
+									<th></th>
+								<? } ?>
+								<th>
+									<?php if($type != MSG_DELETED)
+										echo '<a data-ajax="false" href="'.site_url().'/pm/delete/'.$messages[$i][TF_PM_ID].'/'.$type.'"> x </a>';
+									else
+										echo '<a data-ajax="false" href="'.site_url().'/pm/restore/'.$messages[$i][TF_PM_ID].'"> o </a>'; ?>
+								</th>
+							<?}?>
 						</tr>
 
 						<?php endfor;?>
 					</tbody>
 				</table>
 			<?php else:?>
-				<h1>No messages found.</h1>
+				<h3>No messages found.</h3>
 			<?php endif;?>
 		</div>
 	</div>
